requirements_additional.txt
apex==0.9.10dev
einops==0.8.1
joblib==1.4.2
local_attention==1.10.0
loguru==0.7.2
matplotlib
networkx==3.2.1
numpy<2
ogb==1.3.6
pandas==2.2.3
performer_pytorch
rdkit==2024.3.5
scikit_learn==1.6.1
scipy
seaborn
setuptools==75.8.0
tensorboardX==2.6.2.2
torchmetrics
tqdm==4.66.6
yacs==0.1.8
wandb==0.19.6
